+++ site/src/primitive.rs
@@ -190,9 +190,7 @@ fn all_fills() -> impl IntoView {
             { fill_row(Rows, "Fills row shapes", "⬚0≡⇡ [4 7 3]") }
             { fill_row(Each, "Fills row shapes", "⬚0∵⇡ [3_2 2_4]") }
             { fill_row(Partition, "Fills row shapes", "⬚@ ⊜∘ ≠@ . \"Hey there\"") }
-            { fill_row(Partition, "Sets initial value", "⬚\"- \"⊜⊂ ≠@ . \"Hello world!\"") }
             { fill_row(Group, "Fills row shapes", "⬚0⊕∘ ◿3. [1 8 4 9 3 8 2]") }
-            { fill_row(Group, "Sets initial value", "⬚[]⊕⊂ ◿3. [1 8 4 9 3 8 2]") }
             { fill_row_impl(view!(<Prims prims=[Un, Pop]/>), "Get fill value", "⬚5°◌") }
         </table>
     }