+++ src/primitive/defs.rs
@@ -2007,10 +2007,8 @@ primitive!(
     /// Rows in the second array will be grouped into buckets by the indices in the first array.
     /// Keys `less than``0` will be omitted.
     /// The function then processes each group in order. The result depends on what the function is.
-    /// If the function takes 0 or 1 arguments, then [group] behaves like [rows]. This is called *iterating* [group].
+    /// If the function takes 0 or 1 arguments, then [group] behaves like [rows].
     /// ex: ⊕∘ [0 2 2 1 0 1] [1 2 3 4 5 6]
-    /// If the function takes 2 or more arguments, then [group] behaves like [reduce]. This is called *reducing* [group].
-    /// ex: ⊕⊂ [0 2 2 1 0 1] [1 2 3 4 5 6]
     /// If the values returned by the function do not have the same [shape], concatenation will fail.
     /// ex! ⊕∘ [0 1 0 2 1 1] [1 2 3 4 5 6]
     /// It is common to use [box] to encapsulate groups of different [shape]s.
@@ -2019,6 +2017,12 @@ primitive!(
     /// When combined with [classify], you can do things like counting the number of occurrences of each character in a string.
     /// ex: $ Count the characters in this string
     ///   : ⊟∩≡□ ⊕⊃⊢⧻ ⊛.
+    ///
+    /// If the function takes more than 1 argument, groups are extracted for each argument.
+    /// This example combines each character with the index of its first appearance in the string as well as the number of times it appears.
+    /// ex: ⊕{⊃∩⊢⧻} ⊛ ⟜°⊏ "mississippi"
+    /// Note that multiple values can be returned in this way instead of combining them inside the function.
+    /// ex: ⊕⊃∩⊢⧻ ⊛ ⟜°⊏ "mississippi"
     ///
     /// The indices may be multidimensional.
     /// ex: ⊕□ [0_2 2_1] ["ab" "cd"]
@@ -2040,10 +2044,8 @@ primitive!(
     /// Consecutive rows in the second array that line up with groups of the same key in the markers will be grouped together.
     /// Keys `less or equal``0` will be omitted.
     /// The function then processes each group in order. The result depends on what the function is.
-    /// If the function takes 0 or 1 arguments, then [partition] behaves like [rows]. This is called *iterating* [partition].
+    /// If the function takes 0 or 1 arguments, then [partition] behaves like [rows].
     /// ex: ⊜∘ [0 0 2 2 1 1 3 3] [1 2 3 4 5 6 7 8]
-    /// If the function takes 2 or more arguments, then [partition] behaves like [reduce]. This is called *reducing* [partition].
-    /// ex: ⊜⊂ [0 0 2 2 1 1 3 3] [1 2 3 4 5 6 7 8]
     /// If the values returned by the function do not have the same [shape], concatenation will fail.
     /// ex! ⊜∘ [0 2 3 3 3 0 1 1] [1 2 3 4 5 6 7 8]
     /// It is common to use [box] to encapsulate groups of different [shape]s.
@@ -2055,6 +2057,12 @@ primitive!(
     /// ex: $ 1 1 2 3
     ///   : $ 5 8 13 21
     ///   : ⊜(⊜⋕≠@ .)≠@\n.
+    ///
+    /// If the function takes more than 1 argument, groups are extracted for each argument.
+    /// This example couples each word from a string with its start index.
+    /// ex: ⊜{⊢⊙∘} ≠@  ⟜°⊏ "Hello, how are you?"
+    /// Note that multiple values can be returned in this way instead of combining them inside the function.
+    /// ex: ⊜⊓⊢□ ≠@  ⟜°⊏ "Hello, how are you?"
     ///
     /// [partition] also works with multidimensional markers. Groups are formed from markers that are adjacent along any axis.
     /// Each group will be flattened before being passed to the function.